+++ src/osgUtil/DriveManipulator.cpp
@@ -370,7 +370,7 @@ bool DriveManipulator::calcMovement()
     // return if less then two events have been added.
     if (_ga_t0.get()==NULL || _ga_t1.get()==NULL) return false;
 
-    float dt = _ga_t0->time()-_ga_t1->time();
+    double dt = _ga_t0->time()-_ga_t1->time();
 
     if (dt<0.0f)
     {
